pyproject.toml
[project]
name = "core-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.8",
    "djangorestframework>=3.16.1",
    "gunicorn>=23.0.0",
    "pillow>=12.0.0",
    "psycopg[binary]>=3.2.13",
    "pysonar>=1.2.1.3951",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
